+++ engine/scene/mesh/Bone.h
@@ -13,19 +13,19 @@ namespace coral
     struct KeyPosition
     {
         glm::vec3 position;
-        float timeStamp;
+        double timeStamp;
     };
 
     struct KeyRotation
     {
         glm::quat orientation;
-        float timeStamp;
+        double timeStamp;
     };
 
     struct KeyScale
     {
         glm::vec3 scale;
-        float timeStamp;
+        double timeStamp;
     };
 
     class Bone : public Object
@@ -38,7 +38,7 @@ namespace coral
 
         /* Interpolates b/w positions,rotations & scaling keys based on the curren time of the 
     animation and prepares the local transformation matrix by combining all keys tranformations */
-        void update(float animationTime);
+        void update(double animationTime);
 
         glm::mat4 getLocalTransform();
         std::string getBoneName() const;
@@ -46,31 +46,31 @@ namespace coral
 
         /* Gets the current index on mKeyPositions to interpolate to based on the current 
     animation time */
-        size_t getPositionIndex(float animationTime);
+        size_t getPositionIndex(double animationTime);
 
         /* Gets the current index on mKeyRotations to interpolate to based on the current 
     animation time */
-        size_t getRotationIndex(float animationTime);
+        size_t getRotationIndex(double animationTime);
 
         /* Gets the current index on mKeyScalings to interpolate to based on the current 
     animation time */
-        size_t getScaleIndex(float animationTime);
+        size_t getScaleIndex(double animationTime);
 
     private:
         /* Gets normalized value for Lerp & Slerp*/
-        float getScaleFactor(float lastTimeStamp, float nextTimeStamp, float animationTime);
+        double getScaleFactor(double lastTimeStamp, double nextTimeStamp, double animationTime);
 
         /* figures out which position keys to interpolate b/w and performs the interpolation 
     and returns the translation matrix */
-        glm::mat4 interpolatePosition(float animationTime);
+        glm::mat4 interpolatePosition(double animationTime);
 
         /* figures out which rotations keys to interpolate b/w and performs the interpolation 
     and returns the rotation matrix */
-        glm::mat4 interpolateRotation(float animationTime);
+        glm::mat4 interpolateRotation(double animationTime);
 
         /* figures out which scaling keys to interpolate b/w and performs the interpolation 
     and returns the scale matrix */
-        glm::mat4 interpolateScaling(float animationTime);
+        glm::mat4 interpolateScaling(double animationTime);
 
     private:
         std::vector<KeyPosition> positions;
